design idea: make app list bar size apply to single-window scenarios
this commit is just a design idea, but I was curious how it would look. the issue I have with the current design is that the dots under apps with a single window open were too small to see in small panel scenarios. the change here is to use `bar_size` for single-window apps, and `bar_size * 2.0` for multi-window.

Signed-off-by: Ryan Brue <ryanbrue.dev@gmail.com>